SystemUI: remove mStatusBarContainer to fix focus handling
Since we're killing sysui when restarting themes, there's no need
to try and be graceful about removing views, which is likely why
this extra parent view was added.

This fixes long pressing on a notification bar passing through and
firing off the click intent when the menu shows up.

Change-Id: I3f2539ee100f52a51fc270c34cde39cd312715be
Signed-off-by: Roman Birg <romanbirg@gmail.com>

Conflicts:

	packages/SystemUI/src/com/android/systemui/SystemUI.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java